+++ src/validations/utils/eraseWhiteSpace.js
