+++ src/validations/utils/extractCarName.js
